+++ source/production/F0.Templates/templates/Roslyn-Generator/Roslyn.Generator.UnitTests/Diagnostics/DebugAssertException.cs
@@ -1,0 +1,18 @@
+//-:cnd:noEmit
+#if NETFRAMEWORK
+using System.Diagnostics.CodeAnalysis;
+
+namespace Roslyn.Generator.UnitTests.Diagnostics
+{
+	[SuppressMessage("Design", "CA1032:Implement standard exception constructors", Justification = "Debug.Assert on .NET Framework")]
+	[SuppressMessage("Design", "CA1064:Exceptions should be public", Justification = "Debug.Assert on .NET Framework")]
+	internal sealed class DebugAssertException : Exception
+	{
+		public DebugAssertException(string? message)
+			: base(message)
+		{
+		}
+	}
+}
+#endif
+//+:cnd:noEmit